+++ dto/product.go
@@ -8,7 +8,7 @@ type InsertProductValidator struct {
 	Name      string  `json:"name,omitempty" validate:"required"`
 	UnitPrice float64 `json:"unitPrice,omitempty" validate:"required"`
 	Packaging string  `json:"packaging,omitempty" validate:"required"`
-	Stock     int8    `json:"stock,omitempty" validate:"required"`
+	Stock     float64 `json:"stock,omitempty" validate:"required"`
 	Note      string  `json:"note,omitempty"`
 	CompanyID string  `json:"companyId" validate:"required,uuid4"`
 	CreatedBy uuid.UUID
@@ -19,7 +19,7 @@ type UpdateProductValidator struct {
 	Name      string  `json:"name,omitempty" validate:"required"`
 	UnitPrice float64 `json:"unitPrice,omitempty" validate:"required"`
 	Packaging string  `json:"packaging,omitempty" validate:"required"`
-	Stock     int8    `json:"stock,omitempty" validate:"required"`
+	Stock     float64 `json:"stock,omitempty" validate:"required"`
 	Note      string  `json:"note,omitempty"`
 	CompanyID string  `json:"companyId" validate:"required,uuid4"`
 	UpdatedBy uuid.UUID
@@ -29,7 +29,7 @@ type InsertProduct struct {
 	Name      string
 	UnitPrice float64
 	Packaging string
-	Stock     int8
+	Stock     float64
 	Note      string
 	CompanyID uuid.UUID
 	CreatedBy uuid.UUID
@@ -40,7 +40,7 @@ type UpdateProduct struct {
 	Name      string
 	UnitPrice float64
 	Packaging string
-	Stock     int8
+	Stock     float64
 	Note      string
 	CompanyID uuid.UUID
 	UpdatedBy uuid.UUID
